Use component as plugin for bookmarks
